+++ lib/flex_color_scheme.dart
@@ -55,6 +55,9 @@ export 'src/flex_scheme_on_colors.dart';
 /// FlexColorScheme.
 export 'src/flex_scheme_surface_colors.dart';
 
+/// Enum used to select ink effect in a FlexSubThemesData configuration.
+export 'src/flex_splash_type.dart';
+
 /// An iOS style rounded "squircle" rectangle, also known as super-ellipses.
 export 'src/flex_squircle.dart';